+++ Toolbox/tensor.h
@@ -122,6 +122,9 @@ namespace transform {
     mat compute_jacobian_nominal_to_principal(const mat&);
     mat compute_jacobian_principal_to_nominal(const mat&);
 
+    mat eigen_to_tensor_base(const mat&);
+    mat eigen_to_tensile_derivative(const vec&, const mat&);
+
     template<typename T> Mat<T> skew_symm(const Mat<T>& R) {
         suanpan_assert([&] { if(R.n_elem != 3) throw invalid_argument("need 3 element vector"); });